move types to new manifest package
